// src/stores/keyboard/type.ts
/*
 * @Author: sugar th14200143@163.com
 * @Date: 2022-10-13 15:41:02
 * @LastEditors: sugar th14200143@163.com
 * @LastEditTime: 2022-10-13 15:41:42
 * @FilePath: \ppt-model\src\stores\keyboard\type.ts
 * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
 */
export interface KeyboardState {
  ctrlKeyState: boolean;
  shiftKeyState: boolean;
  spaceKeyState: boolean;
}